#3-task copmlite

+++ 1-basics-of-software-code-development/1/src/main.java
@@ -7,9 +7,9 @@ public class main {
     public static void main(String[] args) throws IOException {
         double a,b,c,z;
         BufferedReader reader = new BufferedReader(new InputStreamReader(System.in));
-        a = Integer.parseInt(reader.readLine());
-        b = Integer.parseInt(reader.readLine());
-        c = Integer.parseInt(reader.readLine());
+        a = Double.parseDouble(reader.readLine());
+        b = Double.parseDouble(reader.readLine());
+        c = Double.parseDouble(reader.readLine());
         z = (((a-3)*b)/2)+c;
         System.out.println(z);
     }
